% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier-package.R
\docType{package}
\name{carrier-package}
\alias{carrier}
\alias{carrier-package}
\title{carrier: Isolate Functions for Remote Execution}
\description{
Sending functions to remote processes can be wasteful of resources because they carry their environments with them. With the carrier package, it is easy to create functions that are isolated from their environment. These isolated functions, also called crates, print at the console with their total size and can be easily tested locally before being sent to a remote.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/carrier}
  \item Report bugs at \url{https://github.com/r-lib/carrier/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
